Fetch The Specified Report Formatting Information

Returns a list of the default font format for the report element

Usage

fetch_report_format(obnd, format_name = "default", verbose = TRUE)

Arguments

obnd

onbrand report object

format_name

Name of report format to fetch; this is defined in the md_def

verbose

Boolean variable when set to TRUE (default) messages will be displayed on the terminal; Messages will be included in the returned list. section for the given report type ("default")

Value

list containing the following elements

isgood: Boolean variable indicating success or failure

msgs: Vector of messages

format_details: List containing the format details for the specified format_name

Examples

obnd = read_template(
       template = file.path(system.file(package="onbrand"), "templates", "report.pptx"),
        mapping = file.path(system.file(package="onbrand"), "templates", "report.yaml"))

fr = fetch_report_format(obnd)