Parse Markdown into Officer as_paragraph Result

Used to take small markdown chunks and return the as_paragraph results. This function will take the markdown specified in str, calls md_to_officer, evals the as_paragraph field from the first paragraph returned, evals that result and returns the object from the as_paragraph command.

md_to_oo(strs, default_format = NULL)

Arguments

strs
vector of strings containing Markdown can contain the following elements:

default_format
list containing the default format for elements not defined with markdown default values (format the same as md_to_officer, default is NULL)

Value

list with the following elements

isgood: Boolean value indicating the result of the function call

msgs: sequence of strings contianing a description of any problems

as_par_cmd:as_paragraph generated code from md_to_officer

oo: as_paragraph officer object resulting from running the as_par_cmd code

Examples

res = md_to_oo("Be **bold**")