onbrand: officer Abstraction Layer for Organizational Templates

The onbrand package creates an abstraction layer that is easily configurable with a yaml file to allow for creation of reproducible reporting work flows across Word and PowerPoint templates.

See also

Author

Maintainer: John Harrold john.m.harrold@gmail.com (ORCID)

Authors:

Bryan Smith r.bryan.smith@gmail.com