Generate Annotated Layout for Report Templates

Elements of slide masters are identified by placeholder labels. As PowerPoint masters are created the labels can be difficult to predict. Word documents are identified by style names. This function will create a layout file identifying all of the elements of each slide master for a PowerPoint template or each paragraph and table style for a Word template.

view_layout(
  template = file.path(system.file(package = "onbrand"), "templates", "report.pptx"),
  output_file = NULL,
  verbose = TRUE
)

Arguments

template
Name of PowerPoint or Word file to annotate (defaults to included PoerPoint template)

output_file
name of file to place the annotated layout information, set to NULL and it will generate a file named layout with the appropriate extension

verbose
Boolean variable when set to TRUE (default) messages will be

Value

List with the following elements

isgood Boolean variable indicating success or failure

rpt Officer with the annotated layout

msgs Vector of messages

Examples

view_layout(template = file.path(system.file(package="onbrand"), "templates", "report.pptx"),
           output_file   = file.path(tempdir(), "layout.pptx"))

#> --------------------------------
#> Generating annotated layout for a report template
#> Template:         /private/var/folders/9x/kwnn6qj93c122hwdlp5hrrnc0000gn/T/RtmpXyWVHd/temp_libpathfc4a73f488da/onbrand/templates/report.pptx
#> Annotated layout: /var/folders/9x/kwnn6qj93c122hwdlp5hrrnc0000gn/T//RtmpN5tvXL/layout.pptx
#> --------------------------------

#> $isgood
#> [1] TRUE
#> 
#> $rpt
#> pptx document with 11 slide(s)
#> Available layouts and their associated master(s) are:
#>                            layout       master
#> 1                     title_slide Office Theme
#> 2                      title_only Office Theme
#> 3                           blank Office Theme
#> 4                    content_text Office Theme
#> 5                    content_list Office Theme
#> 6                   section_slide Office Theme
#> 7         two_content_header_list Office Theme
#> 8         two_content_header_text Office Theme
#> 9  two_content_header_text_backup Office Theme
#> 10               two_content_list Office Theme
#> 11               two_content_text Office Theme
#> 
#> $msgs
#> [1] "--------------------------------"                                                                                                            
#> [2] "Generating annotated layout for a report template"                                                                                           
#> [3] "Template:         /private/var/folders/9x/kwnn6qj93c122hwdlp5hrrnc0000gn/T/RtmpXyWVHd/temp_libpathfc4a73f488da/onbrand/templates/report.pptx"
#> [4] "Annotated layout: /var/folders/9x/kwnn6qj93c122hwdlp5hrrnc0000gn/T//RtmpN5tvXL/layout.pptx"                                                  
#> [5] "--------------------------------"                                                                                                            
#> 

view_layout(template = file.path(system.file(package="onbrand"), "templates", "report.docx"),
           output_file   = file.path(tempdir(), "layout.docx"))

#> --------------------------------
#> Generating annotated layout for a report template
#> Template:         /private/var/folders/9x/kwnn6qj93c122hwdlp5hrrnc0000gn/T/RtmpXyWVHd/temp_libpathfc4a73f488da/onbrand/templates/report.docx
#> Annotated layout: /var/folders/9x/kwnn6qj93c122hwdlp5hrrnc0000gn/T//RtmpN5tvXL/layout.docx
#> --------------------------------

#> $isgood
#> [1] TRUE
#> 
#> $rpt
#> rdocx document with 31 element(s)
#> 
#> * styles:
#>                 Normal              heading 1              heading 2 
#>            "paragraph"            "paragraph"            "paragraph" 
#>              heading 3 Default Paragraph Font           Normal Table 
#>            "paragraph"            "character"                "table" 
#>                No List                Strong1               centered 
#>            "numbering"            "character"            "paragraph" 
#>         table_template    Light List Accent 2         Heading 1 Char 
#>                "table"                "table"            "character" 
#>         Heading 2 Char         Heading 3 Char          graphic title 
#>            "character"            "character"            "paragraph" 
#>            table title     Table Professional                  toc 1 
#>            "paragraph"                "table"            "paragraph" 
#>                  toc 2           Balloon Text      Balloon Text Char 
#>            "paragraph"            "paragraph"            "character" 
#>           reference_id                 header            Header Char 
#>            "character"            "paragraph"            "character" 
#>                 footer            Footer Char             Table Grid 
#>            "paragraph"            "character"                "table" 
#>            page number                   Code 
#>            "character"            "paragraph" 
#> 
#> * Content at cursor location:
#>      row_id is_header cell_id   text col_span row_span style_name content_type
#> 1.1       1      TRUE       1 Number        1        1 Table Grid   table cell
#> 1.3       2     FALSE       1      1        1        1 Table Grid   table cell
#> 1.5       3     FALSE       1      2        1        1 Table Grid   table cell
#> 1.7       4     FALSE       1      3        1        1 Table Grid   table cell
#> 1.9       5     FALSE       1      4        1        1 Table Grid   table cell
#> 2.2       1      TRUE       2   Text        1        1 Table Grid   table cell
#> 2.4       2     FALSE       2   Here        1        1 Table Grid   table cell
#> 2.6       3     FALSE       2   Here        1        1 Table Grid   table cell
#> 2.8       4     FALSE       2   Here        1        1 Table Grid   table cell
#> 2.10      5     FALSE       2   Here        1        1 Table Grid   table cell
#> 
#> $msgs
#> [1] "--------------------------------"                                                                                                            
#> [2] "Generating annotated layout for a report template"                                                                                           
#> [3] "Template:         /private/var/folders/9x/kwnn6qj93c122hwdlp5hrrnc0000gn/T/RtmpXyWVHd/temp_libpathfc4a73f488da/onbrand/templates/report.docx"
#> [4] "Annotated layout: /var/folders/9x/kwnn6qj93c122hwdlp5hrrnc0000gn/T//RtmpN5tvXL/layout.docx"                                                  
#> [5] "--------------------------------"                                                                                                            
#>